Move Java classes into better packages
