+++ pom.xml
@@ -3,7 +3,7 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
 	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>ivy.blockchain.demo</groupId>
+	<groupId>ch.ivy.blockchain.demo</groupId>
 	<artifactId>ivy-blockchain-demo</artifactId>
 	<version>1.0.0-SNAPSHOT</version>
 	<packaging>iar</packaging>
@@ -43,7 +43,7 @@
               <include>**/*.sol</include>
             </includes>
           </soliditySourceFiles>
-          <packageName>ivy.blockchain.demo.model</packageName>
+          <packageName>com.axonivy.demo.blockchain</packageName>
           <sourceDestination>src</sourceDestination>
         </configuration>
       </plugin>
