+++ src/com/axonivy/demo/blockchain/SimpleCounter.java
@@ -1,4 +1,4 @@
-package ivy.blockchain.demo.model;
+package com.axonivy.demo.blockchain;
 
 import java.math.BigInteger;
 import java.util.Arrays;
@@ -24,7 +24,7 @@ import org.web3j.tx.TransactionManager;
  * <p>Generated with web3j version 3.0.2.
  */
 public final class SimpleCounter extends Contract {
-    private static final String BINARY = "60606040526005600055341561001157fe5b5b60bd806100206000396000f300606060405263ffffffff60e060020a6000350416634f2be91f811460345780636deebae31460435780638ada066e146052575bfe5b3415603b57fe5b60416071565b005b3415604a57fe5b6041607d565b005b3415605957fe5b605f608a565b60408051918252519081900360200190f35b6000805460010190555b565b600080546000190190555b565b6000545b905600a165627a7a723058200e2516a412dec2f436233a5f2f13322e0a6be632fcb68f6293bd4cd866342e7a0029";
+    private static final String BINARY = "60606040526005600055341561001457600080fd5b60e4806100226000396000f30060606040526004361060525763ffffffff7c01000000000000000000000000000000000000000000000000000000006000350416634f2be91f811460575780636deebae31460695780638ada066e146079575b600080fd5b3415606157600080fd5b6067609b565b005b3415607357600080fd5b606760a6565b3415608357600080fd5b608960b2565b60405190815260200160405180910390f35b600080546001019055565b60008054600019019055565b600054905600a165627a7a723058209a0a6901ae367bce39bc40ad52010875da67a5eaa54e76482631ecbdf92fb5040029";
 
     private SimpleCounter(String contractAddress, Web3j web3j, Credentials credentials, BigInteger gasPrice, BigInteger gasLimit) {
         super(BINARY, contractAddress, web3j, credentials, gasPrice, gasLimit);
